requirements.txt
PyYAML==3.11
argparse==1.2.1
gnureadline==6.3.3
netCDF4==1.1.3
numpy==1.9.1
pandas==0.15.2
python-dateutil==2.4.0
pytz==2014.10
requests==2.5.1
